Make separate entropy tests
